Use a single message signal and track whether connection succeeded
Previously there was just an error signal with the intention to add a
separate signal for each type of event. This could get kind of awkward
to use so it has just been replaced with a single ‘message‘ signal
which has an enum to describe what happened. The connection now also
detects when we've successfully connected so that it can send a
different message depending on whether the error was before or after
we connected.