Handle the version message
The messages are now read from the peers and validated. A table of
command names and handlers is added which currently just has the
version command. The details of the version command are bundled up
into a message struct and emitted to the ntb_network. The ntb_network
validates it and sends a verack message.