Reject messages that are too old or created in the future
There is now a new type of object called rejected for hashes that we
received that we don't think are valid. Objects that are older than
2.5 days or created in the future are currently given this type.
Anything with a proof-of-work that is too low should also be placed
here although that is not implemented yet.

It now also doesn't keep blobs in memory that are older than 10
minutes. If a peer requests one of these then they can be loaded from
disk.