Handle getdata requests
The connection now has a queue of pending blobs to send. These are
copied to the output buffer just before writing, up to 1kb. The
ntb_network uses this to respond to getdata requests. It can also
queue a hash without a blob in which case the ntb_connection will ask
the store for the object and send it when it is received. It will only
load one blob at a time to avoid keeping too many blobs in memory.