Accept pubkey objects for up to 4 weeks
The official client keeps pubkey objects for this long so if we don't
do the same we'll get lots of annoying warnings when clients send us a
tonne of old keys.

To make this work the blob type enum has been moved to ntb-proto and a
function is added to determine the age given the type. An additional 6
hours is added to each of the ages when deciding whether to garbage
collect old objects so that we won't repeatedly ask for objects from
peers with dodgy clocks.

The type enum in ntb_network_inventory has been split into a state and
a type so that it can reuse the same type enum values as the rest of
the code. The separate lists for each of the object types have been
combined into one because all of the lists were always used together
anyway.