Add a command line option to add peer addresses and use only them
This adds an option like -p 'x.x.x.x:xxxx' to add to the list of
initial addrs that the network will try to connect to. It also adds
the '-e' option to prevent it from connecting to any peers apart from
those listed with the -p options. This is mainly for useful for
debugging so that you can force it to just connect to a local peer.