Save the address list
The address list is now saved at most every ten minutes whenever a new
address is added. The list is saved in a text file via the ntb_store
in the data directory. The store's filename buf now only contains the
root data directory without the /objects suffix so that it can also be
used to generate the addr list filename.