Simplify the division for calculating the POW target
The code for dividing 2⁶⁴ by a value was way over the top considering
that the dividend is so close to a 64-bit value. We can instead just
divide 2⁶³ by the value and multiply the result by two. The old code
wouldn't cope with divisors greater than or equal to 2³² which mean it
would get a broken target value for messages greater than about 12MB
for the default work values.