Add code for generating keys
This adds ntb_crypto which represents a thread that will be used for
doing all of the work involving time-consuming cryptography. The tasks
have a cancellable cookie like ntb_store and ntb_pow. Currently it can
only generate a key pair. The key pair is stored in a new
ref-countable struct called ntb_key.