Respond to ping requests and send pongs at regular intervals
The ntb_connection will now send a pong command somewhere between
every 5 and 8 minutes if no other data has been sent. It will also
respond to a ping command by sending an extra pong command. If a peer
doesn't send any data after between 10 and 13 minutes then it will be
assumed to be dead and the socket will be closed. The inaccuracy with
the timing is because both conditions are only checked every 3
minutes.