Add an object to represent the keyring
The plan is that this will keep track of all of the known private keys
and listen for new messages from the network. It will then ask the
crypto thread to try decrypting them. Currently it just has a list of
the keys and has a cancellable function to generate a new one.