Add extra fields to the ntb_key struct
The ntb_key struct now contains all of the fields stored in the
keys.dat file in the official client. These include the stream number,
version, enabled state, decoy state, the proof-of-work difficulty
settings and the last time the public key was advertised. The label is
initialised on construction and everything else is set to default
values.