connection: Only parse the common fields for object commands
I don't think it makes much sense to parse the entire command in
ntb_connection because it won't be able to parse everything anyway
considering that parts of some of the objects are encrypted. Instead
it will be better to leave it up to ntb_crypto to decode the objects.
This way the ntb_network can happily forward on objects that it
doesn't understand such as those with an unsupported version. This
will fix problems handling v1 pubkeys which notbit was previously
rejecting without putting them in the reject list. That was causing it
to request them from every peer that it connects to.