Avoid using the overloaded term ‘message’
Previously the commands retrieved from peers were inconsistently
referred to as either messages or commands. The events emitted by the
connection were also referred to as messages. This patch changes them
to be consistently called commands and events respectively. The only
remaining use of the term ‘message’ now refers to messages contained
in the msg command.