Decrypt and verify messages
The keyring now listens for msg commands and passes them on to the
ntb_crypto instance to try decrypting it with all of the keys. If the
decryption succeeds then the contents will be passed back as an
ntb_blob. The keyring will process the contents and post the
acknowledgement if it is valid. It doesn't currently store the actual
message anywhere.