Separate the counts for incoming and outgoing connections
Previously the network would try to keep up a minimum of 8 total
connections. There is a bit of a vulnerability in that because it
would be trivial for someone to connect to this node 8 times and
prevent it from connecting to anyone else. Then it would own the node
and could start probing the addresses it owns by sending it messages
and checking whether it sends the acknowledgement. The counts for
incoming and outgoing connections are now separated so that regardless
of how many incoming connections there are it will still try to
connect to 8 peers of its own choosing. This somewhat mitigates the
problem but it probably isn't a complete solution if the rogue nodes
were to start advertising a lot of rogue addrs as that would make the
outgoing connections be dodgy as well.

The maximum number of incoming connections has now been greatly
reduced to 8 instead of 128 because 128 seems like too many in
practice.