Fix an assert when garbage collecting rejected inventories
The inv->type field is not filled in for rejected inventories. When
they were being garbage collected this uninitialised type was being used
to determine the maximum age. The ntb_proto_get_max_age_for_type
function aborts in that case. This patch makes it always use the time
for the MSG type for rejected messages.