Default to a log file in $XDG_DATA_HOME when daemonizing
Using stdout doesn't make sense when daemonizing because the file is
closed after forking. Instead it now makes a file in the store
directory. There is a new function to query the store for this
directory. The store is now initialised before the log file is set so
that it can be sure that the directory has been created. This also
means that the store directory will now be created before dropping
priveleges but I don't think this really matters.