main-context: Use an idle handler for fds that don't support epoll
If a file descriptor is added which doesn't support epoll then it will
now install an idle handler which will report that the fd is always
ready for reading and writing. This is the same behaviour that you
would get with regular poll. This is intended to simplify supporting
regular files sent via the IPC mechanism as they don't support the
poll interface.