store: Add a function to iterate all of the stored pubkey blobs
The keyring will eventually need to iterate all of the stored pubkey
objects in order to maintain a hash table indexed by the ripe or tag.
This will be used to determine if a pubkey that we need is in the
network. The function is similar to ntb_store_for_each_blob except
that it loads the entire contents of the blob instead of just the
meta-data.