keyring: Maintain a hash table of pubkey objects indexed by ripe/tag
When the keyring sends a message to a new key it will need to check
key is already in the network before requesting it. In order to do
this the keyring now keeps track of a hash table of pubkey objects
indexed by either the ripe or the tag depending on the pubkey version.
Only the hash of the object is stored so that the keyring will have to
ask the network or the store for the actual keys when it needs them.