Store other people's public keys in the keyring too
The keyring can now contain ntb_keys without a private part too. There
is a function called ntb_key_has_private to detect these keys. These
are also stored in the store but the public keys are stored instead of
the private keys and some of the fields aren't stored.