Add a keyring function to send a message
The ntb_keyring now keeps track of a list of pending messages. A
message can be added to the list with ntb_keyring_send_message. The
source address and a list of destination addresses are required. The
function will report an error if it doesn't have a private key for the
source address. Multiple destination addresses can be sent which will
cause multiple messages to be sent with the same contents.

If the keyring already has the public key for the destination it will
preceed to encrypt the message and post it. Otherwise it will
currently just drop the message on the floor. This needs to be fixed
so that it will query the network for the public key.

The content of the message is saved to the store and loaded again as
needed. However the meta data for the messages is not preserved so if
Notbit quits before the message is sent then it will be forgotten.

The keyring can recognise an acknowledgement and it will remove the
message from its list when it is acknowledged. However it will never
try to resend the message if no acknowledgement is received.