network: Add a function to query whether the ntb_network has an object
This adds ntb_network_get_object which returns a blob given its hash
if it is in the network's hash table. Even if the object is in the
hash table the network may have moved the object out to storage
instead of keeping it in memory so there is an enum to be able to
return this case.