keyring: Request public keys if they are not already known
Instead of dropping an outgoing message on the floor when the public
key is not immediately available it will now try to fetch the key's
object from the ntb_network. If this doesn't work it will initiate a
getpubkey request. Whenever a new pubkey object is encountered it is
checked against the pending messages to see whether this is a response
to our getpubkey request.