Don't dereference an unitialised pointer in ntb_container_of
The previous implementation was dereferencing the sample pointer in
order to get the offset to subtract from the member pointer. The
resulting value is then only used to get a pointer to the member in
order to calculate the offset so it doesn't actually read from the
memory location and shouldn't cause any problems. However this is
probably technically invalid and could have undefined behaviour. It
looks like clang takes advantage of this undefined behaviour and
doesn't actually offset the pointer. It also generates a warning when
it does this.

This patch splits the ntb_container_of macro into two
implementations. Previously the macro was always used in the list
iterator macros like this:

SomeType *sample = ntb_container_of(list_node, sample, link)

Instead of doing that there is now a new macro called
ntb_list_set_iterator which explicitly assigns to the sample pointer
with an initial value before assigning to it again with the real
offset. This redundant initialisation gets optimised out by compiler.

The second macro is still called ntb_container_of but instead of
taking a sample pointer it just directly takes the type name. That way
it can use the standard offsetof macro.