keyring: Split the last send time into two; for getpubkey and msg
Instead of having one last_send_time whose meaning depends on the
state, there are now two; one to keep track of the last time for the
getpubkey request and the other for the msg. This will make it easier
to save the outgoing messages to the store because it won't need to
save the state enum.