network: Restrict the maximum number of rejected inventories
The rejected inventories are an optimisation to potentially reduce the amount of
traffic we request. However, adding something to Notbit's rejected
list doesn't require POW so it could be used as a DOS-attack vector.
In order to mitigate this problem the maximum size of the rejection
list is now limited to 16384 entries.