Use a struct with optional parameters for the key construction funcs
The functions to create keys had too many arguments which was making
them difficult to use. Instead there is now only one ntb_key_new
function which takes a struct for the arguments. The struct has a
flags parameter to specify which arguments are given so that the
unspecified ones can revert to default values. The functions in
ntb_cryto and ntb_keyring that construct keys have also been changed
to use this struct and pass it down to ntb_key.