Don't set the umask when daemonizing
Setting the umask is meant to be done so that the daemon can have
complete control over the permissions of any files it creates by
specifying them at open time. However Notbit doesn't actually do this
so it ends up creating world-writable files which is not good. Instead
we'll just leave the umask at the default so that someone could set it
before launching notbit if need be.

Many thanks to Klaus Alexander Seistrup for reporting this issue.

https://github.com/bpeel/notbit/issues/2