Add an option to use a SOCKSv5 proxy
Adds the -r option which can be used to specify a SOCKSv5 proxy. There
is a separate function in ntb_connection to connect via a proxy. This
uses a new ntb_proxy object to manage the proxy connection state. The
ntb_proxy shares the input and output buffers of the ntb_connection
and the read and write functions of ntb_connection divert to ntb_proxy
until the connection is established.