Update to protocol v3
The main changes are that objects are now sent with a single message for
all types and that the submission time has been replaced with an
expiry time. Because the object type is now directly in the message,
the object store no longer prepends it to the files.

The timestamp protocol type has been removed because they are now
always 64-bit.

The signatures for public keys and messages now include the entire
object header apart from the nonce.

The timestamp is sadly no longer included in the message objects so
instead it now just writes out the current system time when saving a
message in maildir format.