network: Use a byte size for the memory cache limit instead of age
Instead of pruning items from the in-memory cache when they reach a
certain age, they are now pruned if the cache reaches a certain size.
Everytime an in-memory item is accessed it is moved to the front of
the list so that only the least-recently accessed ones are removed.

The advantage of this is that it is less likely to fill up the memory.
For version 3 of the protocol the old approach won’t be possible
anyway because it’s not possible to know the age of an object, only
when it will expire.