save-message: Split quoted-printable words at UTF-8 character boundary
When saving a message that encodes to a quoted-printable line that
would be longer than 76 characters, notbit splits it up into multiple
words. Previously the line split could be anywhere including in the
middle of a UTF-8 character sequence. Now it tries to shift the split
point so that it is always at a UTF-8 character boundary so that each
encoded section is always valid UTF-8.

Fixes #16