main-context: Delay removing timer sources while iterating
In order to iterate the timer sources, they are now all collected into
a temporary list and then marked as busy. If a busy source is removed,
the actual removal from the list is delayed until the iteration is
finished. That way the context can handle any timeout source being
removed during iteration instead of just the current one.

This should fix a problem where a timeout for receiving data on an
ntb_connection causes its own timeout source to be removed as well as
the ntb_network removing its reconnect timer. Having both sources be
removed in the timeout callback was causing the iteration code to
segfault.

Fixes #18