main-context: Fix reinserting timer sources
When the timer sources are dispatched, they are collected into a
temporary list so that they can be removed in any order while
dispatching. The code to reinsert the sources back into the buckets was
broken in that it wasn’t removing the nodes from the temporary list
before reinserting them. That meant that in some cases the next/prev
pointers for surrounding nodes could still point to reinserted node and
the process could get stuck in an infinite loop.

Many thanks to yshurik on Github for tracking this problem down and
creating a detailed bug report.

Fixes #24